Allow custom device id chooser function to be provided as prop
When zero devices matches the wanted `facingMode`, the first video device
found is chosen. That might not always be the wanted behaviour.

These changes allows a custom function to be provided in the `chooseDeviceId`
prop on the `<QrReader>` component. That function will allowed to choose
which device should be used of the possibly matched video devices or all the
video devices discovered.

The default behaviour stays as before, so this has no effect on projects
that doesn't provide the `chooseDeviceId` prop.